==> notebooks/NB3_perform_spectral_counting_NSAF_detected.ipynb <==
# Perform spectral counting (NSAF) on detected peptides
    Author: Anima Sutradhar
    Project: Peptide detectability prediction to improve protein identification in mass spectrometry using machine learning.

# import libraries
import numpy as np
import pandas as pd
import re
import csv

# set display options
#pd.set_option("display.max_rows", None, "display.max_columns", None)

# import datasets
fasta_peptides = pd.read_table('../data/fasta_peptides.tsv')
detected_peptides = pd.read_table('../data/detected_peptides.tsv')

fasta_peptides.shape

detected_peptides.shape

# sort detected_peptides by PEP
detected_peptides = detected_peptides.sort_values(by=['PEP']).reset_index(drop = True)

# remove any peptides in detected_peptides that map to more than one different protein
detected_peptides = detected_peptides.groupby('Peptide').filter(lambda x: x['Protein'].nunique() == 1)
detected_peptides.shape
# it seems no peptides are shared by more than one different protein, as expected (since we removed ambiguous protein groups)

detected_peptides.describe(include="all")

fasta_peptides.describe(include="all")

## Perform spectral counting on detected_peptides

fasta_protein_len_dict = dict(zip(fasta_peptides.Protein, fasta_peptides.Length))

# map protein sequence lengths to each protein in detected_peptides
detected_peptides['Protein_length'] = detected_peptides['Protein'].map(fasta_protein_len_dict)
detected_peptides

# check for any NaN values in "Length" column
detected_peptides.isnull().sum()

# assign a count value of '1' for each PSM (peptide spectrum match)
detected_peptides["PSM"] = 1
detected_peptides

# count number of times each PSM per protein appears
detected_peptides_abundance = detected_peptides.groupby("Protein")["PSM"].sum()
detected_peptides_abundance

# map protein counts to detected_peptides
detected_peptides_abundance = pd.merge(detected_peptides, detected_peptides_abundance, on='Protein', how='left',
                                 suffixes=(None, '_per_protein'), indicator=True) \
                          .query("_merge == 'both'") \
                          .drop('_merge', 1) \
                          .drop('PSM', 1)
detected_peptides_abundance

# calculate spectral count (using NSAF approach) for detected_peptides
detected_peptides_abundance["Quantification"] = detected_peptides_abundance["PSM_per_protein"] / detected_peptides_abundance["Protein_length"]
detected_peptides_abundance

# check if any detected peptides are not in fasta_peptides (should be 0)
len(detected_peptides_abundance[~detected_peptides_abundance["Peptide"].isin(fasta_peptides["Peptide"])])

detected_peptides_abundance[~detected_peptides_abundance["Peptide"].isin(fasta_peptides["Peptide"])]
# after a random spot check these peptides seem to be present in fasta peptides, but a Methionine amino acid comes before it

detected_peptides_abundance.describe(include="all")

# drop any duplicate peptides (i.e. several instances of same peptides that map to same protein) but keep first occurrence
detected_peptides_abundance = detected_peptides_abundance.drop_duplicates(subset = ['Peptide', 'Protein'], 
                                                                          keep = "first").reset_index(drop = True)
detected_peptides_abundance.describe(include="all")

detected_peptides_abundance.shape

# export dataset
detected_peptides_abundance.to_csv("../data/detected_peptides_NSAF.tsv", sep='\t', index=False)
