==> notebooks/NB5_dataset_prep_for_ML.ipynb <==
# Dataset preparation for machine learning
    Author: Anima Sutradhar
    Project: Peptide detectability prediction to improve protein identification in mass spectrometry using machine learning.

# import libraries
import numpy as np
import pandas as pd
import csv

# set display options
#pd.set_option("display.max_rows", None, "display.max_columns", None)

# import datasets
detected_peptides = pd.read_table('../data/detected_peptides_NSAF.tsv')
undetected_peptides = pd.read_table('../data/undetected_peptides_NSAF.tsv')

detected_peptides.head()

undetected_peptides.head()

# take the mean of the entire PEP column in detected peptides, and apply it to all rows in undetected peptides
undetected_peptides["PEP"] = [detected_peptides["PEP"].mean()]*undetected_peptides.shape[0]
undetected_peptides

# drop "PSM_per_protein"
detected_peptides = detected_peptides.drop(labels=['PSM_per_protein'], axis=1)
detected_peptides.head()

# make all column names the same for both datasets
undetected_peptides = undetected_peptides.rename(columns={"Length": "Protein_length"})
undetected_peptides.head()

detected_peptides.head()

# make column order identical for both datasets
undetected_peptides = undetected_peptides[['Protein', 'Peptide', 'PEP', 'Protein_length', 'Quantification']]
undetected_peptides.head()

# export datasets as TSV
detected_peptides.to_csv("../data/detected_peptides_all.tsv", sep='\t', index=False)
undetected_peptides.to_csv("../data/undetected_peptides_all.tsv", sep='\t', index=False)

print(detected_peptides.shape)
print(undetected_peptides.shape)

# load AAIndex1 datasets
aaindex1_df = pd.read_table('../data/aaindex1_df')
detected_peptides_all = pd.read_table('../data/detected_peptides_all_aaindex1.tsv')
undetected_peptides_all = pd.read_table('../data/undetected_peptides_all_aaindex1.tsv')

print(aaindex1_df.shape)
print(detected_peptides_all.shape)
print(undetected_peptides_all.shape)

undetected_peptides_all.head()
